+++ flinkboot-core/src/main/java/io/github/sekelenao/flinkboot/core/internal/execution/customizer/RestartStrategyCustomizer.java
@@ -54,66 +54,61 @@ public final class RestartStrategyCustomizer implements EnvironmentCustomizer {
 
     private void applyFixedDelay(FixedDelayRestartProperties config) {
         config.attempts().ifPresent(this::applyFixedDelayAttempts);
-        config.delayMs().ifPresent(this::applyFixedDelayDelayMs);
+        config.delay().ifPresent(this::applyFixedDelayDelay);
     }
 
     private void applyFixedDelayAttempts(int attempts) {
         toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_FIXED_DELAY_ATTEMPTS, attempts);
     }
 
-    private void applyFixedDelayDelayMs(long delayMs) {
-        var duration = Duration.ofMillis(delayMs);
-        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_FIXED_DELAY_DELAY, duration);
+    private void applyFixedDelayDelay(Duration delay) {
+        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_FIXED_DELAY_DELAY, delay);
     }
 
     private void applyFailureRate(FailureRateRestartProperties config) {
         config.maxFailuresPerInterval().ifPresent(this::applyFailureRateMaxFailuresPerInterval);
-        config.failureIntervalMs().ifPresent(this::applyFailureRateFailureIntervalMs);
-        config.delayMs().ifPresent(this::applyFailureRateDelayMs);
+        config.failureInterval().ifPresent(this::applyFailureRateFailureInterval);
+        config.delay().ifPresent(this::applyFailureRateDelay);
     }
 
     private void applyFailureRateMaxFailuresPerInterval(int maxFailuresPerInterval) {
         toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_FAILURE_RATE_MAX_FAILURES_PER_INTERVAL, maxFailuresPerInterval);
     }
 
-    private void applyFailureRateFailureIntervalMs(long failureIntervalMs) {
-        var duration = Duration.ofMillis(failureIntervalMs);
-        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_FAILURE_RATE_FAILURE_RATE_INTERVAL, duration);
+    private void applyFailureRateFailureInterval(Duration failureInterval) {
+        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_FAILURE_RATE_FAILURE_RATE_INTERVAL, failureInterval);
     }
 
-    private void applyFailureRateDelayMs(long delayMs) {
-        var duration = Duration.ofMillis(delayMs);
-        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_FAILURE_RATE_DELAY, duration);
+    private void applyFailureRateDelay(Duration delay) {
+        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_FAILURE_RATE_DELAY, delay);
     }
 
     private void applyExponentialDelay(ExponentialDelayRestartProperties config) {
-        config.initialBackoffMs().ifPresent(this::applyExponentialDelayInitialBackoffMs);
-        config.maxBackoffMs().ifPresent(this::applyExponentialDelayMaxBackoffMs);
+        config.initialBackoff().ifPresent(this::applyExponentialDelayInitialBackoff);
+        config.maxBackoff().ifPresent(this::applyExponentialDelayMaxBackoff);
         config.backoffMultiplier().ifPresent(this::applyExponentialDelayBackoffMultiplier);
-        config.resetBackoffThresholdMs().ifPresent(this::applyExponentialDelayResetBackoffThresholdMs);
+        config.resetBackoffThreshold().ifPresent(this::applyExponentialDelayResetBackoffThreshold);
         config.jitterFactor().ifPresent(this::applyExponentialDelayJitterFactor);
     }
 
-    private void applyExponentialDelayInitialBackoffMs(long initialBackoffMs) {
-        var duration = Duration.ofMillis(initialBackoffMs);
-        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_EXPONENTIAL_DELAY_INITIAL_BACKOFF, duration);
+    private void applyExponentialDelayInitialBackoff(Duration initialBackoff) {
+        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_EXPONENTIAL_DELAY_INITIAL_BACKOFF, initialBackoff);
     }
 
-    private void applyExponentialDelayMaxBackoffMs(long maxBackoffMs) {
-        var duration = Duration.ofMillis(maxBackoffMs);
-        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_EXPONENTIAL_DELAY_MAX_BACKOFF, duration);
+    private void applyExponentialDelayMaxBackoff(Duration maxBackoff) {
+        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_EXPONENTIAL_DELAY_MAX_BACKOFF, maxBackoff);
     }
 
     private void applyExponentialDelayBackoffMultiplier(double backoffMultiplier) {
         toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_EXPONENTIAL_DELAY_BACKOFF_MULTIPLIER, backoffMultiplier);
     }
 
-    private void applyExponentialDelayResetBackoffThresholdMs(long resetBackoffThresholdMs) {
-        var duration = Duration.ofMillis(resetBackoffThresholdMs);
-        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_EXPONENTIAL_DELAY_RESET_BACKOFF_THRESHOLD, duration);
+    private void applyExponentialDelayResetBackoffThreshold(Duration resetBackoffThreshold) {
+        toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_EXPONENTIAL_DELAY_RESET_BACKOFF_THRESHOLD, resetBackoffThreshold);
     }
 
     private void applyExponentialDelayJitterFactor(double jitterFactor) {
         toConfigure.set(RestartStrategyOptions.RESTART_STRATEGY_EXPONENTIAL_DELAY_JITTER_FACTOR, jitterFactor);
     }
 }
+